+++ magicclass/_gui/_base.py
@@ -49,7 +49,7 @@ from .mgui_ext import (
     _LabeledWidgetAction,
     mguiLike,
 )
-from .utils import get_parameters, define_callback, callable_to_classes
+from .utils import get_parameters, callable_to_classes
 from ._macro import GuiMacro
 
 from ..utils import (
@@ -818,17 +818,11 @@ class ContainerLikeGui(BaseGui, mguiLike, MutableSequence):
         fld.name = fld.name or name
         action = fld.get_action(self)
 
-        if action.support_value:
+        if action.support_value and fld.record:
             # By default, set value function will be connected to the widget.
-            if fld.record:
-                getvalue = type(fld) is MagicField
-                f = value_widget_callback(self, action, name, getvalue=getvalue)
-                action.changed.connect(f)
-
-            # If the field has callbacks, connect it to the newly generated widget.
-            for callback in fld.callbacks:
-                # funcname = callback.__name__
-                action.changed.connect(define_callback(self, callback))
+            getvalue = type(fld) is MagicField
+            f = value_widget_callback(self, action, name, getvalue=getvalue)
+            action.changed.connect(f)
 
         return action
